backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
httpx>=0.27,<1.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
pypdf>=5.1,<6.0
python-pptx>=1.0,<2.0
pytest>=8.3,<9.0
